requirements.txt
aiofiles==0.6.0
aiohttp==3.7.2
Django==3.1.2
django-multiselectfield==0.1.12
django-summernote==0.8.11.6
django-user-agents==0.4.0
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2-binary==2.8.6
requests==2.25.1
ujson==4.0.1
urllib3==1.22
user-agents==2.2.0
whitenoise==5.2.0
mysqlclient
python-memcached
facebook-sdk
django-ckeditor